Moving random name/id generatio and is_editable/readonly functions to utils.py
