+++ steps.py
@@ -15,7 +15,6 @@ import os
 import os.path
 import re
 import tempfile
-import random
 
 RESULTS_DIR = 'results/'
 ENV_DIR = 'instances/'
@@ -1194,30 +1193,6 @@ def get_values(fieldname):
         return map(lambda x : x.text, select.all_selected_options)
     else:
         return []
-
-
-def random_id():
-    return random.randint(10000,99999)
-
-def random_name():
-
-    sylabs = [ "ba", "bu", "zo", "ga", "to", "le", "mu", "ti", "ma", "xo"]
-    name = str(random_id())
-    name = ''.join([ sylabs[int(char)] for char in name ])
-
-    return name
-
-def is_readonly(content):
-
-    classlist = content.get_attribute("class").split()
-    disabled = content.get_attribute("disabled")
-    return "readonlyfield" in classlist and disabled == "true"
-
-def is_editable(content):
-
-    classlist = content.get_attribute("class").split()
-    disabled = content.get_attribute("disabled")
-    return "readonlyfield" not in classlist and disabled != "true"
 
 
 @step('I should see "([^"]*)" as readonly')
